+++ modules/camera_iphone/camera_module.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  in_app_store.h                                                       */
+/*  camera_module.h                                                      */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,55 +28,5 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef STOREKIT_ENABLED
-
-#ifndef IN_APP_STORE_H
-#define IN_APP_STORE_H
-
-#include "core/object.h"
-
-#ifdef __OBJC__
-@class GodotProductsDelegate;
-@class GodotTransactionsObserver;
-
-typedef GodotProductsDelegate InAppStoreProductDelegate;
-typedef GodotTransactionsObserver InAppStoreTransactionObserver;
-#else
-typedef void InAppStoreProductDelegate;
-typedef void InAppStoreTransactionObserver;
-#endif
-
-class InAppStore : public Object {
-
-	GDCLASS(InAppStore, Object);
-
-	static InAppStore *instance;
-	static void _bind_methods();
-
-	List<Variant> pending_events;
-
-	InAppStoreProductDelegate *products_request_delegate;
-	InAppStoreTransactionObserver *transactions_observer;
-
-public:
-	Error request_product_info(Variant p_params);
-	Error restore_purchases();
-	Error purchase(Variant p_params);
-
-	int get_pending_event_count();
-	Variant pop_pending_event();
-	void finish_transaction(String product_id);
-	void set_auto_finish_transaction(bool b);
-
-	void _post_event(Variant p_event);
-	void _record_purchase(String product_id);
-
-	static InAppStore *get_singleton();
-
-	InAppStore();
-	~InAppStore();
-};
-
-#endif
-
-#endif
+void register_camera_types();
+void unregister_camera_types();
