+++ platform/iphone/SCsub
@@ -9,9 +9,6 @@ iphone_lib = [
     "main.m",
     "app_delegate.mm",
     "view_controller.mm",
-    "game_center.mm",
-    "in_app_store.mm",
-    "icloud.mm",
     "ios.mm",
     "joypad_iphone.mm",
     "godot_view.mm",
