+++ platform/iphone/detect.py
@@ -33,9 +33,6 @@ def get_opts():
             "/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain",
         ),
         ("IPHONESDK", "Path to the iPhone SDK", ""),
-        BoolVariable("game_center", "Support for game center", True),
-        BoolVariable("store_kit", "Support for in-app store", True),
-        BoolVariable("icloud", "Support for iCloud", True),
         BoolVariable("ios_exceptions", "Enable exceptions", False),
         ("ios_triple", "Triple for ios toolchain", ""),
     ]
@@ -204,18 +201,6 @@ def configure(env):
             "ARKit",
         ]
     )
-
-    # Feature options
-    if env["game_center"]:
-        env.Append(CPPDEFINES=["GAME_CENTER_ENABLED"])
-        env.Append(LINKFLAGS=["-framework", "GameKit"])
-
-    if env["store_kit"]:
-        env.Append(CPPDEFINES=["STOREKIT_ENABLED"])
-        env.Append(LINKFLAGS=["-framework", "StoreKit"])
-
-    if env["icloud"]:
-        env.Append(CPPDEFINES=["ICLOUD_ENABLED"])
 
     env.Prepend(
         CPPPATH=[
