+++ platform/iphone/os_iphone.h
@@ -38,9 +38,6 @@
 #include "drivers/unix/os_unix.h"
 #include "joypad_iphone.h"
 
-#include "game_center.h"
-#include "icloud.h"
-#include "in_app_store.h"
 #include "ios.h"
 #include "main/input_default.h"
 #include "servers/audio_server.h"
@@ -57,15 +54,6 @@ private:
 
 	AudioDriverCoreAudio audio_driver;
 
-#ifdef GAME_CENTER_ENABLED
-	GameCenter *game_center;
-#endif
-#ifdef STOREKIT_ENABLED
-	InAppStore *store_kit;
-#endif
-#ifdef ICLOUD_ENABLED
-	ICloud *icloud;
-#endif
 	iOS *ios;
 
 	JoypadIPhone *joypad_iphone;
